Rename serializers: Invariant -> Static, Person->DynamicPerson, PersonActivity -> DynamicActivity
