+++ params-composite-foreign-key.ini
@@ -1,5 +1,5 @@
 ldbc.snb.datagen.generator.scaleFactor:snb.interactive.1
 
-ldbc.snb.datagen.serializer.personSerializer:ldbc.snb.datagen.serializer.snb.csv.compositemergeforeign.CSVCompositeMergeForeignPersonSerializer
-ldbc.snb.datagen.serializer.invariantSerializer:ldbc.snb.datagen.serializer.snb.csv.compositemergeforeign.CSVCompositeMergeForeignInvariantSerializer
-ldbc.snb.datagen.serializer.personActivitySerializer:ldbc.snb.datagen.serializer.snb.csv.compositemergeforeign.CSVCompositeMergeForeignPersonActivitySerializer
+ldbc.snb.datagen.serializer.dynamicPersonSerializer:ldbc.snb.datagen.serializer.snb.csv.compositemergeforeign.CSVCompositeMergeForeignDynamicPersonSerializer
+ldbc.snb.datagen.serializer.staticSerializer:ldbc.snb.datagen.serializer.snb.csv.compositemergeforeign.CSVCompositeMergeForeignStaticSerializer
+ldbc.snb.datagen.serializer.dynamicActivitySerializer:ldbc.snb.datagen.serializer.snb.csv.compositemergeforeign.CSVCompositeMergeForeignDynamicActivitySerializer
