+++ params-csv.ini
@@ -1,8 +1,8 @@
 ldbc.snb.datagen.generator.scaleFactor:snb.interactive.0.1
 
-ldbc.snb.datagen.serializer.personSerializer:ldbc.snb.datagen.serializer.snb.csv.basic.CSVPersonSerializer
-ldbc.snb.datagen.serializer.invariantSerializer:ldbc.snb.datagen.serializer.snb.csv.basic.CSVInvariantSerializer
-ldbc.snb.datagen.serializer.personActivitySerializer:ldbc.snb.datagen.serializer.snb.csv.basic.CSVPersonActivitySerializer
+ldbc.snb.datagen.serializer.dynamicPersonSerializer:ldbc.snb.datagen.serializer.snb.csv.basic.CSVDynamicPersonSerializer
+ldbc.snb.datagen.serializer.staticSerializer:ldbc.snb.datagen.serializer.snb.csv.basic.CSVStaticSerializer
+ldbc.snb.datagen.serializer.dynamicActivitySerializer:ldbc.snb.datagen.serializer.snb.csv.basic.CSVDynamicActivitySerializer
 
 # To generate RFC 3339-compliant timestamps (https://tools.ietf.org/html/rfc3339), uncomment the following line:
 #ldbc.snb.datagen.serializer.formatter.StringDateFormatter.dateTimeFormat:yyyy-MM-dd'T'HH:mm:ss.SSS+00:00
