+++ params-foreign-key.ini
@@ -1,5 +1,5 @@
 ldbc.snb.datagen.generator.scaleFactor:snb.interactive.1
 
-ldbc.snb.datagen.serializer.personSerializer:ldbc.snb.datagen.serializer.snb.csv.mergeforeign.CSVMergeForeignPersonSerializer
-ldbc.snb.datagen.serializer.invariantSerializer:ldbc.snb.datagen.serializer.snb.csv.mergeforeign.CSVMergeForeignInvariantSerializer
-ldbc.snb.datagen.serializer.personActivitySerializer:ldbc.snb.datagen.serializer.snb.csv.mergeforeign.CSVMergeForeignPersonActivitySerializer
+ldbc.snb.datagen.serializer.dynamicPersonSerializer:ldbc.snb.datagen.serializer.snb.csv.mergeforeign.CSVMergeForeignDynamicPersonSerializer
+ldbc.snb.datagen.serializer.staticSerializer:ldbc.snb.datagen.serializer.snb.csv.mergeforeign.CSVMergeForeignStaticSerializer
+ldbc.snb.datagen.serializer.dynamicActivitySerializer:ldbc.snb.datagen.serializer.snb.csv.mergeforeign.CSVMergeForeignDynamicActivitySerializer
