+++ params-graphalytics.ini
@@ -1,5 +1,5 @@
 ldbc.snb.datagen.generator.scaleFactor:graphalytics.1
 
-ldbc.snb.datagen.serializer.personSerializer:ldbc.snb.datagen.serializer.graphalytics.CSVPersonSerializerExtended
-ldbc.snb.datagen.serializer.invariantSerializer:ldbc.snb.datagen.serializer.empty.EmptyInvariantSerializer
-ldbc.snb.datagen.serializer.personActivitySerializer:ldbc.snb.datagen.serializer.empty.EmptyPersonActivitySerializer
+ldbc.snb.datagen.serializer.dynamicPersonSerializer:ldbc.snb.datagen.serializer.graphalytics.CSVDynamicPersonSerializerExtended
+ldbc.snb.datagen.serializer.staticSerializer:ldbc.snb.datagen.serializer.empty.EmptyStaticSerializer
+ldbc.snb.datagen.serializer.dynamicActivitySerializer:ldbc.snb.datagen.serializer.empty.EmptyDynamicActivitySerializer
