+++ params-ttl.ini
@@ -1,5 +1,5 @@
 ldbc.snb.datagen.generator.scaleFactor:snb.interactive.1
 
-ldbc.snb.datagen.serializer.personSerializer:ldbc.snb.datagen.serializer.snb.turtle.TurtlePersonSerializer
-ldbc.snb.datagen.serializer.invariantSerializer:ldbc.snb.datagen.serializer.snb.turtle.TurtleInvariantSerializer
-ldbc.snb.datagen.serializer.personActivitySerializer:ldbc.snb.datagen.serializer.snb.turtle.TurtlePersonActivitySerializer
+ldbc.snb.datagen.serializer.dynamicPersonSerializer:ldbc.snb.datagen.serializer.snb.turtle.TurtleDynamicPersonSerializer
+ldbc.snb.datagen.serializer.staticSerializer:ldbc.snb.datagen.serializer.snb.turtle.TurtleStaticSerializer
+ldbc.snb.datagen.serializer.dynamicActivitySerializer:ldbc.snb.datagen.serializer.snb.turtle.TurtleDynamicActivitySerializer
