+++ src/main/java/ldbc/snb/datagen/hadoop/HadoopDynamicPersonSerializer.java
@@ -40,7 +40,7 @@ import ldbc.snb.datagen.generator.DatagenParams;
 import ldbc.snb.datagen.generator.LDBCDatagen;
 import ldbc.snb.datagen.objects.Knows;
 import ldbc.snb.datagen.objects.Person;
-import ldbc.snb.datagen.serializer.PersonSerializer;
+import ldbc.snb.datagen.serializer.DynamicPersonSerializer;
 import ldbc.snb.datagen.serializer.UpdateEventSerializer;
 import org.apache.hadoop.conf.Configuration;
 import org.apache.hadoop.fs.FileSystem;
@@ -58,17 +58,17 @@ import java.io.IOException;
 /**
  * Created by aprat on 10/15/14.
  */
-public class HadoopPersonSerializer {
+public class HadoopDynamicPersonSerializer {
 
     private Configuration conf;
 
-    public static class HadoopPersonSerializerReducer extends Reducer<TupleKey, Person, LongWritable, Person> {
+    public static class HadoopDynamicPersonSerializerReducer extends Reducer<TupleKey, Person, LongWritable, Person> {
 
         private int reducerId;
         /**
          * The id of the reducer.
          **/
-        private PersonSerializer personSerializer_;
+        private DynamicPersonSerializer dynamicPersonSerializer_;
         /**
          * The person serializer
          **/
@@ -80,9 +80,9 @@ public class HadoopPersonSerializer {
             reducerId = context.getTaskAttemptID().getTaskID().getId();
             LDBCDatagen.initializeContext(conf);
             try {
-                personSerializer_ = (PersonSerializer) Class
-                        .forName(conf.get("ldbc.snb.datagen.serializer.personSerializer")).newInstance();
-                personSerializer_.initialize(conf, reducerId);
+                dynamicPersonSerializer_ = (DynamicPersonSerializer) Class
+                        .forName(conf.get("ldbc.snb.datagen.serializer.dynamicPersonSerializer")).newInstance();
+                dynamicPersonSerializer_.initialize(conf, reducerId);
                 if (DatagenParams.updateStreams) {
                     updateSerializer_ = new UpdateEventSerializer(conf, DatagenParams.hadoopDir + "/temp_updateStream_person_" + reducerId, reducerId, DatagenParams.numUpdatePartitions);
                 }
@@ -96,10 +96,10 @@ public class HadoopPersonSerializer {
         public void reduce(TupleKey key, Iterable<Person> valueSet, Context context)
                 throws IOException, InterruptedException {
 //			SN.machineId = key.block;
-            personSerializer_.reset();
+            dynamicPersonSerializer_.reset();
             for (Person p : valueSet) {
                 if (p.creationDate() < Dictionaries.dates.getUpdateThreshold() || !DatagenParams.updateStreams) {
-                    personSerializer_.export(p);
+                    dynamicPersonSerializer_.export(p);
                 } else {
                     updateSerializer_.export(p);
                     updateSerializer_.changePartition();
@@ -107,7 +107,7 @@ public class HadoopPersonSerializer {
 
                 for (Knows k : p.knows()) {
                     if (k.creationDate() < Dictionaries.dates.getUpdateThreshold() || !DatagenParams.updateStreams) {
-                        personSerializer_.export(p, k);
+                        dynamicPersonSerializer_.export(p, k);
                     }
                 }
             }
@@ -116,7 +116,7 @@ public class HadoopPersonSerializer {
 
         @Override
         protected void cleanup(Context context) {
-            personSerializer_.close();
+            dynamicPersonSerializer_.close();
             if (DatagenParams.updateStreams) {
                 try {
                     updateSerializer_.close();
@@ -127,7 +127,7 @@ public class HadoopPersonSerializer {
         }
     }
 
-    public HadoopPersonSerializer(Configuration conf) {
+    public HadoopDynamicPersonSerializer(Configuration conf) {
         this.conf = new Configuration(conf);
     }
 
@@ -148,7 +148,7 @@ public class HadoopPersonSerializer {
         job.setOutputValueClass(Person.class);
         job.setJarByClass(HadoopBlockMapper.class);
         //job.setMapperClass(HadoopBlockMapper.class);
-        job.setReducerClass(HadoopPersonSerializerReducer.class);
+        job.setReducerClass(HadoopDynamicPersonSerializerReducer.class);
         job.setNumReduceTasks(numThreads);
         job.setInputFormatClass(SequenceFileInputFormat.class);
         job.setOutputFormatClass(SequenceFileOutputFormat.class);
