+++ src/main/java/ldbc/snb/datagen/serializer/PersonActivityExporter.java
@@ -46,19 +46,19 @@ import java.io.IOException;
  * Created by aprat on 3/01/16.
  */
 public class PersonActivityExporter {
-    protected PersonActivitySerializer personActivitySerializer_;
+    protected DynamicActivitySerializer dynamicActivitySerializer_;
     protected UpdateEventSerializer updateSerializer_;
     protected FactorTable factorTable_;
 
-    public PersonActivityExporter(PersonActivitySerializer personActivitySerializer, UpdateEventSerializer updateEventSerializer, FactorTable factorTable) {
-        this.personActivitySerializer_ = personActivitySerializer;
+    public PersonActivityExporter(DynamicActivitySerializer dynamicActivitySerializer, UpdateEventSerializer updateEventSerializer, FactorTable factorTable) {
+        this.dynamicActivitySerializer_ = dynamicActivitySerializer;
         this.updateSerializer_ = updateEventSerializer;
         this.factorTable_ = factorTable;
     }
 
     public void export(final Forum forum) throws IOException {
         if (forum.creationDate() < Dictionaries.dates.getUpdateThreshold() || !DatagenParams.updateStreams) {
-            personActivitySerializer_.export(forum);
+            dynamicActivitySerializer_.export(forum);
         } else {
             updateSerializer_.export(forum);
         }
@@ -66,7 +66,7 @@ public class PersonActivityExporter {
 
     public void export(final Post post) throws IOException {
         if (post.creationDate() < Dictionaries.dates.getUpdateThreshold() || !DatagenParams.updateStreams) {
-            personActivitySerializer_.export(post);
+            dynamicActivitySerializer_.export(post);
             factorTable_.extractFactors(post);
         } else {
             updateSerializer_.export(post);
@@ -75,7 +75,7 @@ public class PersonActivityExporter {
 
     public void export(final Comment comment) throws IOException {
         if (comment.creationDate() < Dictionaries.dates.getUpdateThreshold() || !DatagenParams.updateStreams) {
-            personActivitySerializer_.export(comment);
+            dynamicActivitySerializer_.export(comment);
             factorTable_.extractFactors(comment);
         } else {
             updateSerializer_.export(comment);
@@ -84,7 +84,7 @@ public class PersonActivityExporter {
 
     public void export(final Photo photo) throws IOException {
         if (photo.creationDate() < Dictionaries.dates.getUpdateThreshold() || !DatagenParams.updateStreams) {
-            personActivitySerializer_.export(photo);
+            dynamicActivitySerializer_.export(photo);
             factorTable_.extractFactors(photo);
         } else {
             updateSerializer_.export(photo);
@@ -93,7 +93,7 @@ public class PersonActivityExporter {
 
     public void export(final ForumMembership member) throws IOException {
         if (member.creationDate() < Dictionaries.dates.getUpdateThreshold() || !DatagenParams.updateStreams) {
-            personActivitySerializer_.export(member);
+            dynamicActivitySerializer_.export(member);
             factorTable_.extractFactors(member);
         } else {
             updateSerializer_.export(member);
@@ -102,7 +102,7 @@ public class PersonActivityExporter {
 
     public void export(final Like like) throws IOException {
         if (like.date < Dictionaries.dates.getUpdateThreshold() || !DatagenParams.updateStreams) {
-            personActivitySerializer_.export(like);
+            dynamicActivitySerializer_.export(like);
             factorTable_.extractFactors(like);
         } else {
             updateSerializer_.export(like);
