+++ src/main/java/ldbc/snb/datagen/serializer/StaticSerializer.java
@@ -35,61 +35,47 @@
  Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.*/
 package ldbc.snb.datagen.serializer;
 
-import ldbc.snb.datagen.dictionary.Dictionaries;
-import ldbc.snb.datagen.objects.Knows;
-import ldbc.snb.datagen.objects.Person;
-import ldbc.snb.datagen.objects.StudyAt;
-import ldbc.snb.datagen.objects.WorkAt;
+import ldbc.snb.datagen.objects.Organization;
+import ldbc.snb.datagen.objects.Place;
+import ldbc.snb.datagen.objects.Tag;
+import ldbc.snb.datagen.objects.TagClass;
 import org.apache.hadoop.conf.Configuration;
 
 import java.io.IOException;
-import java.util.Iterator;
+
 
 /**
- * Created by aprat on 10/15/14.
+ * Created by aprat on 12/17/14.
  */
-abstract public class PersonSerializer {
-
-    public void export(final Person person) {
-
-        serialize(person);
-
-        long universityId = Dictionaries.universities.getUniversityFromLocation(person.universityLocationId());
-        if ((universityId != -1) && (person.classYear() != -1)) {
-            StudyAt studyAt = new StudyAt();
-            studyAt.year = person.classYear();
-            studyAt.user = person.accountId();
-            studyAt.university = universityId;
-            serialize(studyAt);
-        }
-
-        Iterator<Long> it = person.companies().keySet().iterator();
-        while (it.hasNext()) {
-            long companyId = it.next();
-            WorkAt workAt = new WorkAt();
-            workAt.company = companyId;
-            workAt.user = person.accountId();
-            workAt.year = person.companies().get(companyId);
-            serialize(workAt);
-        }
-    }
-
-    public void export(final Person p, final Knows k) {
-        if (p.accountId() < k.to().accountId())
-            serialize(p, k);
-    }
+abstract public class StaticSerializer {
 
     abstract public void reset();
+
+    public void export(final TagClass tagclass) {
+        serialize(tagclass);
+    }
+
+    public void export(final Place place) {
+        serialize(place);
+    }
+
+    public void export(final Organization organization) {
+        serialize(organization);
+    }
+
+    public void export(final Tag tag) {
+        serialize(tag);
+    }
 
     abstract public void initialize(Configuration conf, int reducerId) throws IOException;
 
     abstract public void close();
 
-    abstract protected void serialize(final Person p);
+    abstract protected void serialize(final Place place);
 
-    abstract protected void serialize(final StudyAt studyAt);
+    abstract protected void serialize(final Organization organization);
 
-    abstract protected void serialize(final WorkAt workAt);
+    abstract protected void serialize(final TagClass tagClass);
 
-    abstract protected void serialize(final Person p, final Knows knows);
+    abstract protected void serialize(final Tag tag);
 }
