+++ src/main/java/ldbc/snb/datagen/serializer/empty/EmptyDynamicPersonSerializer.java
@@ -35,56 +35,53 @@
  Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.*/
 package ldbc.snb.datagen.serializer.empty;
 
-import ldbc.snb.datagen.objects.Organization;
-import ldbc.snb.datagen.objects.Place;
-import ldbc.snb.datagen.objects.Tag;
-import ldbc.snb.datagen.objects.TagClass;
-import ldbc.snb.datagen.serializer.InvariantSerializer;
+import ldbc.snb.datagen.objects.Knows;
+import ldbc.snb.datagen.objects.Person;
+import ldbc.snb.datagen.objects.StudyAt;
+import ldbc.snb.datagen.objects.WorkAt;
+import ldbc.snb.datagen.serializer.DynamicPersonSerializer;
 import org.apache.hadoop.conf.Configuration;
+
+import java.io.IOException;
 
 /**
  * Created by aprat on 30/01/15.
  */
-public class EmptyInvariantSerializer extends InvariantSerializer {
+public class EmptyDynamicPersonSerializer extends DynamicPersonSerializer {
 
     @Override
-    public void initialize(Configuration conf, int reducerId) {
+    public void initialize(Configuration conf, int reducerId) throws IOException {
         //Intentionally left empty
     }
 
     @Override
     public void close() {
         //Intentionally left empty
-
     }
 
     @Override
-    protected void serialize(final Place place) {
+    protected void serialize(final Person p) {
         //Intentionally left empty
-
     }
 
     @Override
-    protected void serialize(final Organization organization) {
+    protected void serialize(final StudyAt studyAt) {
         //Intentionally left empty
-
     }
 
     @Override
-    protected void serialize(final TagClass tagClass) {
+    protected void serialize(final WorkAt workAt) {
         //Intentionally left empty
-
     }
 
     @Override
-    protected void serialize(final Tag tag) {
+    protected void serialize(final Person p, final Knows knows) {
         //Intentionally left empty
-
     }
 
     @Override
     public void reset() {
         //Intentionally left empty
-
     }
+
 }
